+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -95,11 +95,11 @@ import org.odk.collect.androidshared.system.IntentLauncherImpl;
 import org.odk.collect.androidshared.utils.ScreenUtils;
 import org.odk.collect.androidshared.utils.SettingsUniqueIdGenerator;
 import org.odk.collect.androidshared.utils.UniqueIdGenerator;
-import org.odk.collect.async.CoroutineTaskRunner;
+import org.odk.collect.async.coroutines.CoroutineTaskRunner;
 import org.odk.collect.async.Scheduler;
 import org.odk.collect.async.SchedulerBuilder;
-import org.odk.collect.async.WorkManagerTaskSpecRunner;
-import org.odk.collect.async.WorkManagerTaskSpecScheduler;
+import org.odk.collect.async.workmanager.WorkManagerTaskSpecRunner;
+import org.odk.collect.async.workmanager.WorkManagerTaskSpecScheduler;
 import org.odk.collect.async.network.ConnectivityProvider;
 import org.odk.collect.async.network.NetworkStateProvider;
 import org.odk.collect.audioclips.AudioPlayerFactory;
